service-workers/service-worker/fetch-response-no-cors.https.html

<!DOCTYPE html>
<title>Service Worker: Tainting of responses fetched via SW.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js?pipe=sub"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
promise_test(function(t) {
    var SCRIPT = 'resources/fetch-response-no-cors-worker.js';
    var SCOPE = SCRIPT;

    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          t.add_cleanup(function() { frame.remove(); });
          var request = new Request(
              get_host_info().HTTPS_REMOTE_ORIGIN, {mode: 'no-cors'}
            );
          return frame.contentWindow.fetch(request);
        })
      .then(function(resp) {
          assert_equals(resp.type, 'opaque');
        });
  }, 'Verify the tainting of responses fetched via SW');
</script>
</body>
